pyproject.toml
[tool.poetry]
name = "dspy-examples"
version = "0.1.0"
description = ""
authors = ["foxgem <jianhgreat@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
dspy-ai = "^2.4.13"
python-dotenv = "^1.0.1"
newspaper4k = {extras = ["all"], version = "^0.9.3.1"}
lxml-html-clean = "^0.2.0"
nltk = "3.8.1"
gradio = "^4.42.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
